+++ webpack.config.js
@@ -1,3 +1,0 @@
-const dev = require('./config/webpack.dev.config');
-const prod = require('./config/webpack.prod.config');
-module.exports = [prod, dev];